readme: Compile with QtCreator on host
When compiling on the host this application is expected to be compiled
with QtCreator, which runs qmake and make. Currently the readme says do
this manually. QtCreator seperates the build directory by target where
the manual make does not. This means compiling with the IDE for the host
keeps the two architectures seperate.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>