Fix issue with camera warning
The app generates popups to alert the user when the camera is
not connected, however, the first popup doesn't show correctly,
and it might make the app unusable.

While trying a few strategies, it became clear that the version
of Qt used has some bugs that cannot be circumvented early on
in the boot process, therefore a different strategy has to be
defined.

The UI has an empty slot in the uppper grid, therefore use it
to show the status of the camera to get around the nasty issue
with the popup.

Signed-off-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>