MainWindow: Allow the user to cancel file selection
If the user opens the load file menu and closes it without making a
selection a warning popup displays stating their selected file does not
exist. We should check the return value of the dialog to ensure a file
has been selected instead. We can protect against the user from entering
a filename that does not exist by selecting a suitable FileMode for the
dialog.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>