MainWindow: Only Capture Image with an ready camera
The application runs a constant stream with the webcam in the
background if one is connected initialised. However if the camera
has not been loaded yet the capture button will attempt to take
from a stream that does not exist yet, causing a the application to crash.

This also means that the user can see the capture image button when
playing a video file, but clicking it will attempt to take from the
webcam instead of the expected video file.

Only displaying the capture image button when the webcam is loaded
protects the user and application from both scenarios.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>